Allow marking fields as internal.
These fields are not shown during registration and are only visible
in the admin list by default.

The use case is creating fields that can be used on badges, etc. but are
not generally user visible/editable. This is required for GUADEC to print
the wifi username and password on badges.