Implement support for 'purchase' type fields
You can specify a purchase with one or more different options. Each
option can have a different cost, and different amount of items
available.

The payment for the purchase can be taken along with the registration
fee, or as a separate, later payment. The purchase choice cannot be
changed once the payment has been received.

The purchases can be paid in any of the configured currencies, although it
must be the same currency as the registration fee. The prices are
converted to the main currency following the configured conversion_rate.

We also now ask for payment on registration updates if there is still
payment due. Before, if a user registered and then closed the
registration/registration_success view without paying, then they
wouldn't be able to get back to that page to actually make a payment.

Purchases are included in the PayPal payment processing, so the user
can see from the PayPal page what items they are buying. The PayPal SKU
is used to preserve this information across the PayPal transaction so
we can then associate the payment with the correct purchases in our DB.

Users can still make donations after the registration fee & purchases
are paid for. This is now treated a bit separately in the code.

The payment admin view (log in as payadmn@regcfp) shows the payment
status for each user's purchase, so you can keep track of any purchases
that have been allocated but not paid for.

This code builds on the existing support for T-shirts and workshops
that was used in previous years.

https://github.com/puiterwijk/regcfp/issues/157