Use supertest-as-promised instead of supertest
It's more or less the same but allows you to chain steps together with
Promises (via the 'bluebird' library) instead of having to nest
callbacks.

Supertest 2.0.0 apparently supports the same thing, but I couldn't get
it to work myself and it would also only work with Node 0.11.0 and above
as it requires the built-in Promise implementation, not the 'bluebird'
library.