Sort and filter the country data
It's normal for the list of countries to be sorted by first letter of
the name, rather than by the country code.

It's also normal for the list of countries to only list countries that
still exist. Previously people could say they were from Yugoslavia, the
USSR, the European Union and various the other countries which have
status 'deleted' or 'reserved'.