Avoid duplicating list of countries many times
If a 'type: country' field had 'options' defined in the config file, the
options field from the config would have all the countries added to
it... every time utils.get_reg_fields() was called. This led to quite a
long list, which in turn made the tests log very noisy.