Allow setting an error to be displayed on the index page
This will be used when there are internal errors in the OpenID Connect
auth provider that we need to report. The message will have to be vague
so we don't risk exposing internal secrets, but a "something failed"
message is better than a 'login' button that simply does nothing, with
no indication that there's an error.